moved "reading logged in storage" into util
